extract elimination forest kernels into separate file
