+++ core/factorization/cholesky_kernels.hpp
@@ -20,13 +20,6 @@ namespace gko {
 namespace kernels {
 
 
-#define GKO_DECLARE_CHOLESKY_COMPUTE_SKELETON_TREE(IndexType)               \
-    void compute_skeleton_tree(std::shared_ptr<const DefaultExecutor> exec, \
-                               const IndexType* row_ptrs,                   \
-                               const IndexType* cols, size_type size,       \
-                               IndexType* out_row_ptrs, IndexType* out_cols)
-
-
 #define GKO_DECLARE_CHOLESKY_SYMBOLIC_COUNT(ValueType, IndexType)        \
     void symbolic_count(                                                 \
         std::shared_ptr<const DefaultExecutor> exec,                     \
@@ -42,13 +35,6 @@ namespace kernels {
         const gko::factorization::elimination_forest<IndexType>& forest, \
         matrix::Csr<ValueType, IndexType>* l_factor,                     \
         const array<IndexType>& tmp_storage)
-
-
-#define GKO_DECLARE_CHOLESKY_FOREST_FROM_FACTOR(ValueType, IndexType) \
-    void forest_from_factor(                                          \
-        std::shared_ptr<const DefaultExecutor> exec,                  \
-        const matrix::Csr<ValueType, IndexType>* factors,             \
-        gko::factorization::elimination_forest<IndexType>& forest)
 
 
 #define GKO_DECLARE_CHOLESKY_INITIALIZE(ValueType, IndexType)                 \
@@ -73,14 +59,10 @@ namespace kernels {
 
 
 #define GKO_DECLARE_ALL_AS_TEMPLATES                               \
-    template <typename IndexType>                                  \
-    GKO_DECLARE_CHOLESKY_COMPUTE_SKELETON_TREE(IndexType);         \
     template <typename ValueType, typename IndexType>              \
     GKO_DECLARE_CHOLESKY_SYMBOLIC_COUNT(ValueType, IndexType);     \
     template <typename ValueType, typename IndexType>              \
     GKO_DECLARE_CHOLESKY_SYMBOLIC_FACTORIZE(ValueType, IndexType); \
-    template <typename ValueType, typename IndexType>              \
-    GKO_DECLARE_CHOLESKY_FOREST_FROM_FACTOR(ValueType, IndexType); \
     template <typename ValueType, typename IndexType>              \
     GKO_DECLARE_CHOLESKY_INITIALIZE(ValueType, IndexType);         \
     template <typename ValueType, typename IndexType>              \
