+++ reference/CMakeLists.txt
@@ -19,6 +19,7 @@ target_sources(ginkgo_reference
     distributed/partition_kernels.cpp
     distributed/vector_kernels.cpp
     factorization/cholesky_kernels.cpp
+    factorization/elimination_forest_kernels.cpp
     factorization/factorization_kernels.cpp
     factorization/ic_kernels.cpp
     factorization/ilu_kernels.cpp
